% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Get a query for a SQLite file}
\usage{
get_query(filename, sql)
}
\arguments{
\item{filename}{SQLite file location}

\item{sql}{String containing the SQL query to be performed}
}
\description{
Used internally by \code{\link{query_sql}}. Use that function instead to access data. The
use of this function is not recommended
}
\seealso{
\code{\link{query_sql}} to perform standard queries of data
}
\keyword{internal}

