% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_solution.R
\name{is_otf_rplexos}
\alias{is_otf_rplexos}
\alias{enable_otf_rplexos}
\alias{disable_otf_rplexos}
\alias{check_otf_rplexos}
\title{Enable or disable on-the-fly mode}
\usage{
is_otf_rplexos()

enable_otf_rplexos(msg = T)

disable_otf_rplexos(msg = T)

check_otf_rplexos()
}
\arguments{
\item{msg}{When true it prints a message}
}
\description{
The on-the-fly mode will only generate the sql db after a specific query. This mode is very
useful for a Plexos db with many report items while only a handful are required for the post-
processing. The initial generation of the db will be very quick. The first queries will be slower
but subsequent queries will be at normal speeds again.
}
