% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{is_sample_stats}
\alias{is_sample_stats}
\title{Test if elements in sample column are statistics}
\usage{
is_sample_stats(x)
}
\arguments{
\item{x}{Vector of sample values from an rplexos query}
}
\description{
In stochastic simulations, PLEXOS will return sample results and their statistics together. This function
makes it easy to separate them with a filter.
}
\examples{
# Process the folder with the solution file provided by rplexos
location <- location_solution_rplexos()
process_folder(location)

# Query data
db <- plexos_open(location)
res <- query_month(db, "Generator", "Generation")
res \%>\% filter(sample_stats(sample))    # To obtain statistics
res \%>\% filter(!sample_stats(sample))   # To obtain sample results
}

