% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{list_folders}
\alias{list_folders}
\title{Get list of folders in the working directory}
\usage{
list_folders()
}
\description{
List of existing folders in the working directory. This function is used when the wildcard symbol (\code{"*"})
is provided to the \code{\link{process_folder}} and \code{\link{plexos_open}} functions.
}
\seealso{
\code{\link{setwd}}, \code{\link{process_folder}}, \code{\link{plexos_open}}
}

