% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{location_solution_rplexos}
\alias{location_solution_rplexos}
\alias{location_input_rplexos}
\title{Shortcut functions for rplexos sample files}
\usage{
location_solution_rplexos(phase = "ST")

location_input_rplexos()
}
\arguments{
\item{phase}{this specifies if it is the ST or the LT solution.}
}
\description{
These functions return the folder where the PLEXOS sample files
are located. They are uses in different examples and the vignettes.
}
