% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plexos_open.R
\name{plexos_open}
\alias{plexos_open}
\title{Open all PLEXOS databases}
\usage{
plexos_open(folders = ".", names = folders)
}
\arguments{
\item{folders}{character. Folder(s) where the data is located (each folder represents a scenario)}

\item{names}{character. Scenario names}
}
\description{
The default for \code{folders} is the working directory. If the wildcard \code{"*"} is provided, all
the folders in the working directory will be processed (the list of folders if provided by
the \code{\link{list_folders}} function).
}
\seealso{
\code{\link{plexos_close}}

\code{\link{query_master}}
}

