% Generated by roxygen2 (4.0.2): do not edit by hand
\name{process_folder}
\alias{process_folder}
\alias{process_solution}
\title{Convert PLEXOS solutions to SQLite databases}
\usage{
process_folder(folders = ".", keep.temp = FALSE)

process_solution(file, keep.temp = FALSE)
}
\arguments{
\item{folders}{Folder(s) to process (See details)}

\item{keep.temp}{Should temporary databases be preserved?}

\item{file}{Single PLEXOS solution file to process}
}
\description{
Functions to process PLEXOS solutions and dump them into SQLite databases that are easy
to read.
}
\details{
\code{process_solution} converts a single zipped PLEXOS solution into a SQLite database.
The resulting database file has a simplified structure so data can be retrieved easily.
The temporary database contains the original data structure in the PLEXOS solution file.

\code{process_folders} is used to process an entire directory. It searches each folder for
PLEXOS solutions and applies \code{process_solution} to each one of them.

The default folder is the working directory. If the wildcard \code{"*"} is provided, all
the folders in the working directory will be processed (the list of folders if provided by
the \code{\link{list_folders}} function).
}
\examples{
\dontrun{process_folder()}
\dontrun{process_folder("HiWind")}
\dontrun{process_solution("HiWind/Model WWSIS_c_RT_CoreB_M01_SC3 Solution.zip")}
}

