% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_input.R
\name{process_input}
\alias{process_input}
\title{Convert PLEXOS input databases to SQLite format}
\usage{
process_input(file)
}
\arguments{
\item{file}{Single PLEXOS input file to process}
}
\description{
Functions to process PLEXOS input databases and dump them into a SQLite file that are easy
to read.
}

