% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_summary.R
\name{query_class}
\alias{query_class}
\alias{query_class_member}
\alias{query_generator}
\alias{query_region}
\alias{query_zone}
\title{Get list of objects from all databases}
\usage{
query_class(db)

query_class_member(db, class)

query_generator(db)

query_region(db)

query_zone(db)
}
\arguments{
\item{db}{PLEXOS database object}

\item{class}{Type of class to query}
}
\description{
Get the list of objects and classes in each database. Shortcuts for generators, regions and zones
are provided for convenience.
}
\seealso{
Other special.queries: \code{\link{query_config}},
  \code{\link{query_log}}, \code{\link{query_phase}},
  \code{\link{query_property}}, \code{\link{query_sql}},
  \code{\link{query_time}}
}

