% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_master_each}
\alias{query_master_each}
\title{Open one database and query data}
\usage{
query_master_each(db, time, col, prop, columns = "name", time.range = NULL,
  filter = NULL, phase = 4)
}
\arguments{
\item{db}{PLEXOS database object}

\item{time}{character. Table to query from (interval, day, week, month, year)}

\item{col}{character. Collection to query}

\item{prop}{character vector. Property or properties to query}

\item{columns}{character. Data columns to query or aggregate by (defaults to \code{name})}

\item{time.range}{POSIXt or character. Range of dates of length 2 (given as date, datetime or character in 'ymdhms' or 'ymd' format)}

\item{filter}{list. Used to filter by data columns (see details)}

\item{phase}{integer. PLEXOS optimization phase (1-LT, 2-PASA, 3-MT, 4-ST)}
}
\description{
Used internally by \code{\link{query_master}}. Use that function instead to access data. The
use of this function is not recommended
}
\seealso{
\code{\link{query_master}} to query PLEXOS solutions
}
\keyword{internal}

