% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query_property}
\alias{query_property}
\title{Get list of available properties}
\usage{
query_property(db)
}
\arguments{
\item{db}{PLEXOS database object}
}
\description{
Produce a list of available properties, their units and the collections they belong to.
Additionally, a column is created for each scenario that indicates in how many databases
the property appears.
}
\seealso{
\code{\link{plexos_open}} to create the PLEXOS database object
}

