% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query_sql}
\alias{query_sql}
\title{Get query for all scenarios}
\usage{
query_sql(db, sql)
}
\arguments{
\item{db}{PLEXOS database object}

\item{sql}{String containing the SQL query to be performed}
}
\description{
Send a SQL query to all the files in a PLEXOS database object.
}
\seealso{
\code{\link{plexos_open}} to create the PLEXOS database object

\code{\link{query_master}} to perform standard queries of data
}

