% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query_summary.R
\name{query_time}
\alias{query_time}
\title{Get time spans from all databases}
\usage{
query_time(db)
}
\arguments{
\item{db}{PLEXOS database object}
}
\description{
Get the time limits and time step lengths for each simulation phase.
}
\seealso{
Other special.queries: \code{\link{query_band}},
  \code{\link{query_phase}}, \code{\link{query_sample}},
  \code{\link{query_timeslice}}; \code{\link{query_class}},
  \code{\link{query_class_member}},
  \code{\link{query_generator}},
  \code{\link{query_region}}, \code{\link{query_zone}};
  \code{\link{query_config}}; \code{\link{query_log}},
  \code{\link{query_log_steps}};
  \code{\link{query_property}}; \code{\link{query_sql}}
}

