% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/debug.R
\name{start_debug_rplexos}
\alias{check_debug_rplexos}
\alias{is_debug_rplexos}
\alias{start_debug_rplexos}
\alias{stop_debug_rplexos}
\title{Enable or disable debug mode}
\usage{
start_debug_rplexos()

stop_debug_rplexos()

check_debug_rplexos()

is_debug_rplexos()
}
\description{
The debug mode will print progress on screen and additional information to
help diagnose problems.
}

