% Generated by roxygen2 (4.0.2): do not edit by hand
\name{valid_columns}
\alias{valid_columns}
\title{Get list of valid columns}
\usage{
valid_columns()
}
\description{
List of valid columns accepted in \code{\link{query_master}}, \code{\link{sum_master}} and related functions.
}
\seealso{
\code{\link{query_master}}, \code{\link{sum_master}}
}

