% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{location_solution_rplexos}
\alias{location_input_rplexos}
\alias{location_solution_rplexos}
\title{Shortcut functions for rplexos sample files}
\usage{
location_solution_rplexos()

location_input_rplexos()
}
\description{
These functions return the folder where the PLEXOS sample files
are located. They are uses in different examples and the vignettes.
}

