% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plexos_open.R
\name{plexos_close}
\alias{plexos_close}
\title{Close all PLEXOS databases}
\usage{
plexos_close(db)
}
\arguments{
\item{db}{PLEXOS database object}
}
\description{
Close all the open connections to PLEXOS SQLite databases. This function
completely erases the provided object.
}
\seealso{
\code{\link{plexos_open}} to create the PLEXOS database object
}

