% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plexos_open}
\alias{plexos_open}
\title{Open all PLEXOS databases}
\usage{
plexos_open(folders = ".", names = folders)
}
\arguments{
\item{folders}{Character of folder name(s) where the data is located (each folder represents a scenario)}

\item{names}{Scenario names}
}
\description{
Open all PLEXOS databases
}
\seealso{
\code{\link{plexos_close}}

\code{\link{query_master}}
}

