% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plexos_open.R
\name{plexos_open}
\alias{plexos_open}
\title{Open all PLEXOS databases}
\usage{
plexos_open(folders = ".", names = folders)
}
\arguments{
\item{folders}{character. Folder(s) where the data is located (each folder represents a scenario)}

\item{names}{character. Scenario names}
}
\description{
The default for \code{folders} is the working directory. If the wildcard \code{"*"} is provided, all
the folders in the working directory will be processed (the list of folders if provided by
the \code{\link{list_folders}} function).
}
\details{
Do not rename the SQLite databases created with the \code{\link{process_folder}} family of functions.
The code expects those filenames to remain unchanged.
}
\seealso{
\code{\link{query_master}} to perform standard queries of data

\code{\link{query_sql}} to perform custom queries
}

