% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query_log}
\alias{query_log}
\alias{query_log_steps}
\title{Query log file information}
\usage{
query_log(db)

query_log_steps(db)
}
\arguments{
\item{db}{PLEXOS database object}
}
\description{
During the processing of the PLEXOS databases, information from the log file is saved
into the database. This includes solution times and infeasibilities for the different phases.
}
\seealso{
\code{\link{plexos_open}} to create the PLEXOS database object
}

