% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_summary.R
\name{query_phase}
\alias{query_band}
\alias{query_phase}
\alias{query_sample}
\alias{query_timeslice}
\title{Get summary information from all databases}
\usage{
query_phase(db)

query_sample(db)

query_timeslice(db)

query_band(db)
}
\arguments{
\item{db}{PLEXOS database object}
}
\description{
Get the list of phases, samples, timeslices and bands  that are available in each database.
}
\seealso{
Other special.queries: \code{\link{query_class}},
  \code{\link{query_config}}, \code{\link{query_log}},
  \code{\link{query_property}}, \code{\link{query_sql}},
  \code{\link{query_time}}
}

