% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_summary.R
\name{query_time}
\alias{query_time}
\title{Get time spans from all databases}
\usage{
query_time(db)
}
\arguments{
\item{db}{PLEXOS database object}
}
\description{
Get the time limits and time step lengths for each simulation phase.
}
\seealso{
Other special.queries: \code{\link{query_class}},
  \code{\link{query_config}}, \code{\link{query_log}},
  \code{\link{query_phase}}, \code{\link{query_property}},
  \code{\link{query_sql}}
}

